Clean up and split `BubbledInspectableNode` into its own file.
