+++ packages/breadboard-ui/src/elements/activity-log/activity-log.ts
@@ -6,7 +6,6 @@
 
 import {
   ErrorObject,
-  InspectablePortList,
   InspectableRun,
   InspectableRunEvent,
   InspectableRunNodeEvent,
